Please work

+++ FlightSoftware/TEMT.h
@@ -4,6 +4,6 @@ extern void TEMT_init();
 
 extern void TEMT_run();
 
-extern int get_TEMT_Data();
+extern float get_TEMT_Data();
 
 extern bool TEMT_initialized;
